Highlight first item with closeOnSelect to false
Currently multiselect having closeOnSelect attribute set to false not highlights first item in dropdown selection. It disrupts the workflow, that the operator needs to move the hands from the keyboard, to the mouse, for every address that needs to be written manually. By enabling the highlighting of the first item in dropdown selection, the productivity of the operators will be increased and the workflow will not be disrupted.